ml/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
absl-py==0.11.0
    # via
    #   tensorboard
    #   tensorflow
argon2-cffi==20.1.0
    # via notebook
asgiref==3.3.1
    # via django
astunparse==1.6.3
    # via tensorflow
async-generator==1.10
    # via nbclient
attrs==20.3.0
    # via jsonschema
backcall==0.2.0
    # via ipython
bleach==3.3.0
    # via
    #   nbconvert
    #   readme-renderer
boto3==1.16.63
    # via -r requirements.in
botocore==1.19.63
    # via
    #   boto3
    #   s3transfer
cachetools==4.2.1
    # via google-auth
certifi==2020.12.5
    # via requests
cffi==1.14.5
    # via
    #   argon2-cffi
    #   cryptography
chardet==4.0.0
    # via requests
colorama==0.4.4
    # via twine
contextlib2==0.6.0.post1
    # via tensorflow-object-detection-api
cryptography==3.4.5
    # via secretstorage
cycler==0.10.0
    # via matplotlib
cython==0.29.21
    # via tensorflow-object-detection-api
decorator==4.4.2
    # via ipython
defusedxml==0.6.0
    # via nbconvert
django-storages==1.11.1
    # via -r requirements.in
django==3.1.6
    # via
    #   -r requirements.in
    #   django-storages
    #   djangorestframework
djangorestframework==3.12.2
    # via -r requirements.in
docutils==0.16
    # via readme-renderer
entrypoints==0.3
    # via nbconvert
gast==0.3.3
    # via tensorflow
google-auth-oauthlib==0.4.2
    # via tensorboard
google-auth==1.26.1
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.35.0
    # via
    #   tensorboard
    #   tensorflow
h5py==2.10.0
    # via tensorflow
idna==2.10
    # via requests
ipykernel==5.4.3
    # via
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   notebook
    #   qtconsole
ipython-genutils==0.2.0
    # via
    #   nbformat
    #   notebook
    #   qtconsole
    #   traitlets
ipython==7.20.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
ipywidgets==7.6.3
    # via jupyter
jedi==0.18.0
    # via ipython
jeepney==0.6.0
    # via
    #   keyring
    #   secretstorage
jinja2==2.11.3
    # via
    #   nbconvert
    #   notebook
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonpickle==2.0.0
    # via -r requirements.in
jsonschema==3.2.0
    # via nbformat
jupyter-client==6.1.11
    # via
    #   ipykernel
    #   jupyter-console
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.2.0
    # via jupyter
jupyter-core==4.7.1
    # via
    #   jupyter-client
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyter==1.0.0
    # via tensorflow-object-detection-api
jupyterlab-pygments==0.1.2
    # via nbconvert
jupyterlab-widgets==1.0.0
    # via ipywidgets
keras-preprocessing==1.1.2
    # via tensorflow
keyring==22.0.1
    # via twine
kiwisolver==1.3.1
    # via matplotlib
lxml==4.6.2
    # via tensorflow-object-detection-api
markdown==3.3.3
    # via tensorboard
markupsafe==1.1.1
    # via jinja2
matplotlib==3.3.4
    # via
    #   -r requirements.in
    #   tensorflow-object-detection-api
mistune==0.8.4
    # via nbconvert
mysqlclient==2.0.3
    # via -r requirements.in
nbclient==0.5.2
    # via nbconvert
nbconvert==6.0.7
    # via
    #   jupyter
    #   notebook
nbformat==5.1.2
    # via
    #   ipywidgets
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.1
    # via nbclient
notebook==6.2.0
    # via
    #   jupyter
    #   widgetsnbextension
# numpy==1.24.1
    # via
    #   -r requirements.in
    #   h5py
    #   keras-preprocessing
    #   matplotlib
    #   opencv-python
    #   opt-einsum
    #   pandas
    #   scipy
    #   tensorboard
    #   tensorflow
oauthlib==3.1.0
    # via requests-oauthlib
opencv-python
# opencv-python==4.5.1.48
    # via -r requirements.in
opt-einsum==3.3.0
    # via tensorflow
packaging==20.9
    # via bleach
pandas==1.2.2
    # via -r requirements.in
pandocfilters==1.4.3
    # via nbconvert
parso==0.8.1
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==8.1.0
    # via
    #   -r requirements.in
    #   matplotlib
    #   tensorflow-object-detection-api
pkginfo==1.7.0
    # via twine
prometheus-client==0.9.0
    # via notebook
prompt-toolkit==3.0.16
    # via
    #   ipython
    #   jupyter-console
protobuf==3.14.0
    # via
    #   tensorboard
    #   tensorflow
    #   tensorflow-object-detection-api
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pycparser==2.20
    # via cffi
pygments==2.8.0
    # via
    #   ipython
    #   jupyter-console
    #   jupyterlab-pygments
    #   nbconvert
    #   qtconsole
    #   readme-renderer
pyparsing==2.4.7
    # via
    #   matplotlib
    #   packaging
pyrsistent==0.17.3
    # via jsonschema
python-dateutil==2.8.1
    # via
    #   botocore
    #   jupyter-client
    #   matplotlib
    #   pandas
pytz==2021.1
    # via
    #   django
    #   pandas
pyzmq==22.0.3
    # via
    #   jupyter-client
    #   notebook
    #   qtconsole
qtconsole==5.0.2
    # via jupyter
qtpy==1.9.0
    # via qtconsole
readme-renderer==28.0
    # via twine
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
requests-toolbelt==0.9.1
    # via twine
requests==2.25.1
    # via
    #   requests-oauthlib
    #   requests-toolbelt
    #   tensorboard
    #   twine
rfc3986==1.4.0
    # via twine
rsa==4.7
    # via google-auth
s3transfer==0.3.4
    # via boto3
scipy
    # via tensorflow
secretstorage==3.3.1
    # via keyring
send2trash==1.5.0
    # via notebook
six==1.15.0
    # via
    #   absl-py
    #   argon2-cffi
    #   astunparse
    #   bleach
    #   cycler
    #   google-auth
    #   google-pasta
    #   grpcio
    #   h5py
    #   jsonschema
    #   keras-preprocessing
    #   protobuf
    #   python-dateutil
    #   readme-renderer
    #   tensorboard
    #   tensorflow
sqlparse==0.4.1
    # via django
tensorboard-plugin-wit==1.8.0
    # via tensorboard
tensorboard
# tensorboard==2.4.1
    # via tensorflow
tensorflow-estimator==2.3.0
    # via tensorflow
# tensorflow-object-detection-api==0.1.1
    # via -r requirements.in
# tensorflow=2.3.0
tensorflow-macos
    # via
    #   -r requirements.in
    #   tensorflow-object-detection-api
termcolor==1.1.0
    # via tensorflow
terminado==0.9.2
    # via notebook
testpath==0.4.4
    # via nbconvert
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   notebook
    #   terminado
tqdm==4.56.2
    # via twine
traitlets==5.0.5
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
twine==3.3.0
    # via tensorflow-object-detection-api
urllib3==1.25.11
    # via
    #   botocore
    #   requests
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
werkzeug==1.0.1
    # via tensorboard
wheel==0.36.2
    # via
    #   astunparse
    #   tensorboard
    #   tensorflow
    #   tensorflow-object-detection-api
widgetsnbextension==3.5.1
    # via ipywidgets
wrapt==1.12.1
    # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
